Redefines `indices_dtype` as two concepts: * `indices_dtype` to index graph pieces (sizes and adjacency indices).   (int32 by default). * `row_splits_dtype` to index ragged tensor partitions.   (int64 by default, as in tf.RaggedTensor).
Adds helpers `with_indices_dtype()` and `with_row_splits_dtype()`.

Ensures indices consistency for all graph tensor pieces and applies auto-casting on construction.

Removed `row_splits_dtype` from `random_graph_tensor`.

Simplified rules for GraphPieceBase metadata updates: each piece may define their
own metadata fields, completely independently. Before the base class tried to automatically align metadata fields for all its children. Now it is not required and not enforced.

Allows to load graph tensors stored with inconsistent indices (for backwards compatibility)

PiperOrigin-RevId: 553073221

+++ tensorflow_gnn/graph/graph_tensor_encode_test.py
@@ -105,7 +105,7 @@ class TestWriteExample(tf.test.TestCase, parameterized.TestCase):
                    if tf.TensorShape(shape).is_fully_defined()
                    else tf.RaggedTensorSpec(shape, dtype))
     spec = create_tensor(tensor_spec)
-    rgraph = gr.random_graph_tensor(spec, row_splits_dtype=tf.int64)
+    rgraph = gr.random_graph_tensor(spec)
     example = ge.write_example(rgraph)
     serialized = tf.constant(example.SerializeToString())
     pgraph = io.parse_single_example(spec, serialized, validate=True)
